Move to a hot-swappable module style of processing
Installing this will only install the APO which will check
a defined path for any pluggable processing DLLs. If found, it'll
use that DLL for processing, else act like a passthrough.

Remove all instances of rnnoise as it will be moved to a separate
repo as a standalone plugin based module.

Also fix VS path readme bruh moment